pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
description = "AI-powered support agent with WhatsApp integration for server infrastructure support"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.0",
    "websockets>=12.0",

    # Voice Processing
    "openai-whisper>=20231117",
    "ffmpeg-python>=0.2.0",

    # LLM
    "ollama>=0.1.0",
    "openai>=1.10.0",

    # SSH for tools
    "asyncssh>=2.14.0",

    # Database
    "aiomysql>=0.2.0",

    # Web UI
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
